refactor terminal
