+++ lib/src/utils/terminal/dialog.rs
@@ -1,6 +1,14 @@
 use crate::{async_command_pipe, print, SPINNERS};
+use dialoguer::{theme::ColorfulTheme, Confirm};
 use anyhow::{anyhow, Result};
 use indicatif::{ProgressBar, ProgressStyle};
+use std::io::Result as IOResult;
+
+pub fn proceed(prompt: &str) -> IOResult<bool> {
+    let color_theme = &ColorfulTheme::default();
+    let dialog = String::from(prompt);
+    Confirm::with_theme(color_theme).with_prompt(dialog).interact()
+}
 
 pub async fn spinner_cmd(cmd: &str, exec_msg: &'static str, finish_msg: &'static str) -> Result<()> {
     if let Some(arrows) = SPINNERS.get("arrows") {
@@ -19,12 +27,33 @@ pub async fn spinner_cmd(cmd: &str, exec_msg: &'static str, finish_msg: &'static
     }
 }
 
+pub fn spinner(exec_msg: &'static str, tick_strings: &[&str]) -> ProgressBar {
+    let spinner = ProgressBar::new_spinner();
+    let spinner_style = ProgressStyle::default_spinner()
+        .template("{msg:.green} {spinner:.green}")
+        .tick_strings(tick_strings);
+    spinner.enable_steady_tick(80);
+    spinner.set_style(spinner_style);
+    spinner.set_message(exec_msg);
+    spinner
+}
+
 #[cfg(test)]
 mod test {
-    // use crate::spinner_cmd;
+    // use super::*;
+    #[tokio::test]
+    #[ignore]
+    async fn test_spinner() {
+        unimplemented!();
+    }
     #[tokio::test]
     #[ignore]
     async fn test_spinner_cmd() {
+        unimplemented!();
+    }
+    #[tokio::test]
+    #[ignore]
+    async fn test_proceed() {
         unimplemented!();
     }
 }
