--- lib/src/utils/terminal/dialog/mod.rs
@@ -1,6 +1,0 @@
-pub mod proceed;
-pub use proceed::proceed;
-pub mod spinner_cmd;
-pub use spinner_cmd::spinner_cmd;
-pub mod spinner;
-pub use spinner::spinner;
