--- lib/src/utils/terminal/shell/change_shell_config.rs
@@ -1,22 +1,0 @@
-use crate::{check_shell_config_env, print, write_shell_config, PATHS};
-use anyhow::Result;
-
-pub async fn change_shell_config() -> Result<()> {
-    print("", "Checking for shell configuration")?;
-    for (key, value) in PATHS.iter() {
-        if !check_shell_config_env(key).await? {
-            write_shell_config(value).await?;
-        }
-    }
-    print("green", "Shell configured")
-}
-
-#[cfg(test)]
-mod test {
-    // use crate::change_shell_config;
-    #[tokio::test]
-    #[ignore]
-    async fn test_change_shell_config() {
-        unimplemented!();
-    }
-}
