--- lib/src/utils/terminal/shell/check_shell.rs
@@ -1,16 +1,0 @@
-use crate::check_env;
-use anyhow::Result;
-
-pub async fn check_shell() -> Result<String> {
-    check_env("SHELL")
-}
-
-#[cfg(test)]
-mod test {
-    // use crate::check_shell;
-    #[tokio::test]
-    #[ignore]
-    async fn test_check_shell() {
-        unimplemented!();
-    }
-}
