--- lib/src/utils/terminal/shell/get_shell_profile_file.rs
@@ -1,7 +1,0 @@
-use crate::{check_env, check_shell, match_shell};
-use anyhow::Result;
-
-pub async fn get_shell_profile_file() -> Result<String> {
-    match_shell(&check_shell().await?)?;
-    check_env("SHELL_PROFILE_FILE")
-}
