--- lib/src/utils/terminal/shell/mod.rs
@@ -1,20 +1,0 @@
-pub mod ask_shell_config;
-pub use ask_shell_config::ask_shell_config;
-pub mod change_shell_config;
-pub use change_shell_config::change_shell_config;
-pub mod match_shell;
-pub use match_shell::match_shell;
-pub mod check_shell;
-pub use check_shell::check_shell;
-pub mod check_shell_config_env;
-pub use check_shell_config_env::check_shell_config_env;
-pub mod setup_shell;
-pub use setup_shell::setup_shell;
-pub mod write_shell_config;
-pub use write_shell_config::write_shell_config;
-pub mod export_shell_variables;
-pub use export_shell_variables::export_shell_variables;
-pub mod source_shell;
-pub use source_shell::source_shell;
-pub mod get_shell_profile_file;
-pub use get_shell_profile_file::get_shell_profile_file;
