--- lib/src/utils/terminal/stdout/mod.rs
@@ -1,6 +1,0 @@
-pub mod print_emoji;
-pub use print_emoji::print_emoji;
-pub mod print;
-pub use print::print;
-pub mod to_color;
-pub use to_color::to_color;
